eliminated CSVDataFrameLoader pass-through class, folded its minimal functionality into CSVDataFrameParser
