+++ core/src/main/java/edu/stanford/futuredata/macrobase/pipeline/PipelineUtils.java
@@ -3,7 +3,7 @@ package edu.stanford.futuredata.macrobase.pipeline;
 import com.fasterxml.jackson.databind.ObjectMapper;
 import edu.stanford.futuredata.macrobase.datamodel.DataFrame;
 import edu.stanford.futuredata.macrobase.datamodel.Schema;
-import edu.stanford.futuredata.macrobase.ingest.CSVDataFrameLoader;
+import edu.stanford.futuredata.macrobase.ingest.CSVDataFrameParser;
 import edu.stanford.futuredata.macrobase.ingest.RESTDataFrameLoader;
 import edu.stanford.futuredata.macrobase.util.MacrobaseException;
 
@@ -31,7 +31,7 @@ public class PipelineUtils {
             List<String> requiredColumns
     ) throws Exception {
         if(inputURI.substring(0, 3).equals("csv")) {
-            CSVDataFrameLoader loader = new CSVDataFrameLoader(inputURI.substring(6), requiredColumns);
+            CSVDataFrameParser loader = new CSVDataFrameParser(inputURI.substring(6), requiredColumns);
             loader.setColumnTypes(colTypes);
             DataFrame df = loader.load();
             return df;
