+++ lib/src/test/java/edu/stanford/futuredata/macrobase/UnsupervisedCSVTest.java
@@ -5,7 +5,7 @@ import edu.stanford.futuredata.macrobase.analysis.summary.fpg.FPGExplanation;
 import edu.stanford.futuredata.macrobase.analysis.summary.fpg.FPGrowthSummarizer;
 import edu.stanford.futuredata.macrobase.datamodel.DataFrame;
 import edu.stanford.futuredata.macrobase.datamodel.Schema;
-import edu.stanford.futuredata.macrobase.ingest.CSVDataFrameLoader;
+import edu.stanford.futuredata.macrobase.ingest.CSVDataFrameParser;
 import edu.stanford.futuredata.macrobase.ingest.DataFrameLoader;
 import org.junit.Before;
 import org.junit.Test;
@@ -33,7 +33,7 @@ public class UnsupervisedCSVTest {
         schema.put("latency", Schema.ColType.DOUBLE);
         schema.put("location", Schema.ColType.STRING);
         schema.put("version", Schema.ColType.STRING);
-        DataFrameLoader loader = new CSVDataFrameLoader(
+        DataFrameLoader loader = new CSVDataFrameParser(
                 "src/test/resources/sample.csv"
         ).setColumnTypes(schema);
         df = loader.load();
